create toastContext and use proivder instead of storage
