+++ app/component/toast.tsx
@@ -1,13 +1,9 @@
+import { useContext, useEffect } from "react"
 import Failed from "react:~assets/failed.svg"
 import Sent from "react:~assets/sent.svg"
 import Success from "react:~assets/success.svg"
 
-export type ToastStatus = "success" | "failed" | "sent"
-
-type ToastProps = {
-  status: ToastStatus
-  message: string
-}
+import { ToastContext, type ToastStatus } from "~app/context/toastContext"
 
 const statusIcon: Record<ToastStatus, JSX.Element> = {
   success: <Success />,
@@ -20,8 +16,21 @@ const statusTextColor: Record<ToastStatus, string> = {
   failed: "text-[#FF9393]",
   sent: "text-[#7EE0FF]"
 }
-export const Toast = (props: ToastProps) => {
-  const { status, message } = props
+export const Toast = () => {
+  const { toast, setToast } = useContext(ToastContext)
+  const { status, message } = toast
+
+  useEffect(() => {
+    if (toast && toast.message && toast.status) {
+      const timer = setTimeout(() => {
+        setToast(null, null)
+      }, 2000)
+      return () => clearTimeout(timer)
+    }
+  }, [toast, setToast])
+
+  if (!status || !message) return
+
   return (
     <div className="absolute w-[358px] top-[32px] flex gap-[8px] items-center bg-black rounded-2xl p-[16px] shadow-[4px_8px_0px_0px_rgba(0,0,0,0.16)]">
       {statusIcon[status]}
