+++ storage/local/state.ts
@@ -1,4 +1,3 @@
-import type { ToastStatus } from "~app/component/toast"
 import { AccountType } from "~packages/account"
 import { EntryPointVersion } from "~packages/bundler"
 import type {
@@ -22,10 +21,6 @@ export type State = {
   }
   pendingTransaction: {
     [txId: string]: TransactionPending
-  }
-  toast: {
-    message: string | null
-    status: ToastStatus | null
   }
 }
